requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pyyaml==6.0.1
pymongo==4.6.1
requests==2.31.0
